requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
databases
alembic
python-dotenv
requests
jinja2
pandas
scikit-learn
matplotlib
langgraph
